show key as appropriate for current keyboard layout
