+++ src/gwt/src/org/rstudio/core/rebind/command/ShortcutsEmitter.java
@@ -19,7 +19,6 @@ import com.google.gwt.core.ext.TreeLogger.Type;
 import com.google.gwt.core.ext.UnableToCompleteException;
 import com.google.gwt.user.rebind.SourceWriter;
 
-import org.rstudio.core.client.Pair;
 import org.rstudio.core.client.command.KeyboardShortcut;
 import org.w3c.dom.Element;
 import org.w3c.dom.Node;
@@ -35,6 +34,13 @@ import java.util.List;
 
 public class ShortcutsEmitter
 {
+   private static class ShortcutKeyCombination
+   {
+      public String key;
+      public String keyCode;
+      public int modifiers;
+   }
+   
    public ShortcutsEmitter(TreeLogger logger,
                            String groupName, 
                            Element shortcutsEl) throws UnableToCompleteException
@@ -112,7 +118,8 @@ public class ShortcutsEmitter
                               String title,
                               String disableModes) throws UnableToCompleteException
    {
-      List<Pair<Integer, String>> keys = new ArrayList<Pair<Integer, String>>();
+      List<ShortcutKeyCombination> keys = new ArrayList<ShortcutKeyCombination>();
+      
       for (String keyCombination : shortcut.split("\\s+"))
       {
          String[] chunks = keyCombination.split("\\+");
@@ -142,20 +149,17 @@ public class ShortcutsEmitter
             }
          }
             
-         // Validate the key name.
-         String key = toKey(chunks[chunks.length - 1]);
-         if (key == null)
-         {
-            logger_.log(Type.ERROR,
-                  "Invalid shortcut '" + shortcut + "', only " +
-                  "modified alphanumeric characters, enter, " +
-                  "left, right, up, down, pageup, pagedown, " +
-                  "and tab are valid");
-            throw new UnableToCompleteException();
-         }
+         // Extract the key name.
+         String key = chunks[chunks.length - 1];
          
-         // Push the parsed keys to the list.
-         keys.add(new Pair<Integer, String>(modifiers, key));
+         // Push the keys to the list.
+         ShortcutKeyCombination combination = new ShortcutKeyCombination();
+         
+         combination.key = key;
+         combination.keyCode = toKeyCode(key);
+         combination.modifiers = modifiers;
+         
+         keys.add(combination);
       }
       
       // Emit the relevant code registering these shortcuts.
@@ -167,11 +171,15 @@ public class ShortcutsEmitter
       
       if (keys.size() == 1)
       {
-         int modifiers = keys.get(0).first;
-         String key = keys.get(0).second;
+         ShortcutKeyCombination combination = keys.get(0);
          writer.println("ShortcutManager.INSTANCE.register(" +
-               modifiers + ", " +
-               key + ", " +
+         
+               // Key set
+               "\"" + combination.key + "\", " +
+               combination.keyCode + ", " +
+               combination.modifiers + ", " +
+               
+               // Command + group
                command + ", " +
                "\"" + shortcutGroup + "\", " +
                "\"" + title + "\", " + 
@@ -179,21 +187,26 @@ public class ShortcutsEmitter
       }
       else if (keys.size() == 2)
       {
-         int m1    = keys.get(0).first;
-         String k1 = keys.get(0).second;
-         int m2    = keys.get(1).first;
-         String k2 = keys.get(1).second;
+         ShortcutKeyCombination c1 = keys.get(0);
+         ShortcutKeyCombination c2 = keys.get(1);
          
          writer.println("ShortcutManager.INSTANCE.register(" +
-               m1 + ", " +
-               k1 + ", " +
-               m2 + ", " +
-               k2 + ", " +
+         
+               // First key set
+               "\"" + c1.key + "\", " +
+               c1.keyCode + ", " +
+               c1.modifiers + ", " +
+               
+               // Second key set
+               "\"" + c2.key + "\", " +
+               c2.keyCode + ", " +
+               c2.modifiers + ", " +
+               
+               // Command + group
                command + ", " +
                "\"" + shortcutGroup + "\", " +
                "\"" + title + "\", " + 
                "\"" + disableModes + "\");");
-         
       }
       else
       {
@@ -210,7 +223,7 @@ public class ShortcutsEmitter
       }
    }
 
-   private String toKey(String val)
+   private String toKeyCode(String val)
    {
       if (val.matches("^[a-zA-Z0-9]$"))
          return "'" + val.toUpperCase() + "'";
