+++ src/gwt/src/org/rstudio/studio/client/workbench/views/source/Source.java
@@ -392,23 +392,10 @@ public class Source implements InsertSourceHandler,
       }
 
       // fake shortcuts for commands which we handle at a lower level
-      commands.goToHelp().setShortcut(new KeyboardShortcut(112));
-      commands.goToDefinition().setShortcut(new KeyboardShortcut(113));
-      commands.codeCompletion().setShortcut(
-                                    new KeyboardShortcut(KeyCodes.KEY_TAB));
+      commands.goToHelp().setShortcut(new KeyboardShortcut("F1", KeyCodes.KEY_F1, KeyboardShortcut.NONE));
+      commands.goToDefinition().setShortcut(new KeyboardShortcut("F2", KeyCodes.KEY_F2, KeyboardShortcut.NONE));
+      commands.codeCompletion().setShortcut(new KeyboardShortcut("Tab", KeyCodes.KEY_TAB, KeyboardShortcut.NONE));
       
-      // See bug 3673 and https://bugs.webkit.org/show_bug.cgi?id=41016
-      if (BrowseCap.isMacintosh())
-      {
-         ShortcutManager.INSTANCE.register(
-               KeyboardShortcut.META | KeyboardShortcut.ALT,
-               192,
-               commands.executeNextChunk(), 
-               "Execute",
-               commands.executeNextChunk().getMenuLabel(false), 
-               "");
-      }
-
       events.addHandler(ShowContentEvent.TYPE, this);
       events.addHandler(ShowDataEvent.TYPE, this);
       events.addHandler(OpenObjectExplorerEvent.TYPE, this);
